Player Plan & Frame Customisation Start
We've started out planning on how the player customisation will start, with them being able to select a size/type of frame and their style of legs, both of which determine the ty[e/number of weapons they can have, their speed, etc.

Additionally, we've started a basic implementation of this with a quick tester script that can toggle the GFX for the subcomponents of a player's frames, along with cleaning up the hierarchy by removing the unused options when a Finalise() function is called.

Note: We've got all the options under the prefabs so that we can give them custom positions per frame/slot/etc. It also means that we only need to remove them, rather than instantiating them when the player selects.
Note 2: We are currently toggling & destroying based on their order in the hierarchy. We should update this to something more robust (E.g. A dictionary with Enum to GameObject Key-Value Pairs)