Mech Customisation Work
We've worked on the Mech Customisation, making it more robust by having the frames, legs, weapons, and ability gfx elements be toggled based on an active data object (ScriptableObject) of their associated type.

This should make it so that we can just pass the active element and the individual GFX elements will handle everything, letting us expand it easier and making it more robust to hierarchy & data order changes.